+++ app/assets/javascripts/app/views/conversations_form_view.js
@@ -8,6 +8,10 @@ app.views.ConversationsForm = Backbone.View.extend({
 
   initialize: function(opts) {
     this.contacts = _.has(opts, "contacts") ? opts.contacts : null;
+    if(!this.contacts || this.contacts.length === 0) {
+      this.displayNoContactsMessage();
+      return;
+    }
     this.prefill = [];
     if (_.has(opts, "prefillName") && _.has(opts, "prefillValue")) {
       this.prefill = [{name : opts.prefillName,
@@ -30,6 +34,12 @@ app.views.ConversationsForm = Backbone.View.extend({
       preFill: this.prefill
     }).focus();
     $("#contact_ids").attr("aria-labelledby", "toLabel");
+  },
+
+  displayNoContactsMessage: function() {
+    $("form#new_conversation").replaceWith(
+      "<div class=\"well text-center\">" + Diaspora.I18n.t("conversation.new.no_contacts") + "</div>"
+    );
   },
 
   keyDown : function(evt) {